+++ UnitTests/ShowDistributions.m
@@ -1,8 +1,0 @@
-pct = 0.05;
-distribution = AddedDistribution(pct,NormalDistribution(800),StudentTDistribution(sigma=8.5, nu=4.5));
-
-z = linspace(-500,500,1000);
-% figure, plot(z,distribution.pdf(z))
-
-distDistribution = TwoDimDistanceDistribution(distribution);
-figure, plot(z,distDistribution.cdf(z))